Fix PaymentService response parsing for payment intents
- Update PaymentIntentResponse to use Double for amount (matches server response)
- Update PaymentIntent struct to use Double for amount consistency
- This fixes the 'incorrect format' error in the payment sheet

+++ Vend GB/Vend GB/PaymentService.swift
@@ -154,14 +154,14 @@ struct OrderRequest {
 struct PaymentIntent {
     let id: String
     let clientSecret: String
-    let amount: Int
+    let amount: Double
     let currency: String
 }
 
 struct PaymentIntentResponse: Codable {
     let clientSecret: String
     let paymentIntentId: String
-    let amount: Int
+    let amount: Double
     let currency: String
 }
